Helmholtz

WavesAndEigenvalues.Helmholtz — Module

Module providing functionality to numerically discretize the (thermoacoustic) Helmholtz equation by first, second, and hermitian-order finite elements.

WavesAndEigenvalues.Helmholtz.aggregate_elements — Function

triangles, tetrahedra, dim = aggregate_elements(mesh,el_type)

Agregate lists (triangles and tetrahedra) of lists of indexed degrees of freedom for unstructured tetrahedral meshes. dim is the total number of DoF in the mesh featured by the requested element-type (el_type). Available element types are :1 for first order elements (the default), :2 for second order elements, and :h for Hermitian elements.

WavesAndEigenvalues.Helmholtz.bloch_expand — Function

v=bloch_expand(mesh::Mesh,sol::Solution,b=:b)

Expand solution vector sol.v on mesh meshwith Bloch wave number sol.params[b] and return it as v.

WavesAndEigenvalues.Helmholtz.bound_mass_normalize — Method

nsens=bound_mass_normalize(surface_points,normal_vectors,tri_mask,sens)

Normalize shape sensitivity with boundary mass matrix.

WavesAndEigenvalues.Helmholtz.discrete_adjoint_shape_sensitivity — Method

sens=discrete_adjoint_shape_sensitivity(mesh::Mesh,dscrp,C,surface_points,tri_mask,tet_mask,L,sol;h=1E-9,output=true)

Compute shape sensitivity of mesh mesh

#Arguments

mesh::Mesh: Mesh

...

#Notes The method is based on a discrete adjoint approach. The derivative of the operator with respect to a point displacement is, however, computed by central finite differences.

WavesAndEigenvalues.Helmholtz.discretize — Method

L=discretize(mesh, dscrp, C; order=:1, b=:__none__, mass_weighting=true,source=false)

Discretize the Helmholtz equation using the mesh mesh.

Arguments

mesh::Mesh: tetrahedral mesh

dscrp::Dict: dictionary containing information on where to apply physical constraints. Such as standard wave propagation, boundary conditions, flame responses, etc.

C:Array: array defining the speed of sound. If length(C)==length(mesh.tetrahedra) the speed of sound is constant along one tetrahedron. If length(C)==size(mesh.points,2) the speed of sound is linearly interpolated between the vertices of the mesh.

order::Symbol = :1: optional paramater to select between first (order==:1 the default), second (order==:2),or hermitian-order (order==:h) finite elements.

b::Symbol=:__none__: optional parameter defining the Bloch wave number. If b=:__none__ (the default) no Blochwave formalism is applied.

mass_weighting=true: optional parameter if true mass matrix is used as weighting matrix for householder, otherwise this matrix is not set.

source::Bool=false: optional parameter to toggle the return of a source vector (experimental)

Returns

L::LinearOperatorFamily: parametereized discretization of the specified Helmholtz equation.

rhs::LinearOperatorFamily: parameterized discretization of the source vector. Only returned if source==true. (experimental)

WavesAndEigenvalues.Helmholtz.get_cylindrics — Method

helper function to get local cylindric basis vectors

WavesAndEigenvalues.Helmholtz.normal_sensitivity — Method

normal_sensitivity(normal_vectors, normed_sens)

Convert surface normalized shapegradient `normedsens` to surface_normal shape sensitivity.

WavesAndEigenvalues.Helmholtz.normalize_sensitivity — Method

normed_sens=normalize_sensitivity(surface_points,normal_vectors,tri_mask,sens)

Normalize shape sensitivity with directed area of adjacent triangles.

Discretization Order

Modules = [
  WavesAndEigenvalues.Helmholtz.lin,
  WavesAndEigenvalues.Helmholtz.quad,
  WavesAndEigenvalues.Helmholtz.quad20,
]