Mesh

WavesAndEigenvalues.Meshutils — Module

Module containing functionality to read and process tetrahedral meshes from gmsh.

WavesAndEigenvalues.Meshutils.Mesh — Type

Definition of the mesh type

Fields

name::String: the name of the mesh.

points::Array: 3×N array containing the coordinates of the N points defining the mesh.

lines::List: List of simplices defining the edges of the mesh

triangles::List: List of simplices defining the surface triangles of the mesh

tetrahedra::List: List of simplices defining the tetrahedra of the mesh

domains::Dict: Dictionairy defining the domains of the mesh. See comments below.

file::String: path to the file containing the mesh.

tri2tet::Array: Array of length length(tetrahedra) containing the indices of the connected tetrahedra.

dos: special field meant to contain symmetry information of highly symmetric meshes.

Notes

The meshes are supposed to be tetrahedral. All simplices (lines, triangles, and tetrahedra) are stored as lists of simplices. Simplices are lists of integers containing the indices of the points (i.e. the column number in the points array) forming the simplex. This means a line is a two-entry list, a triangle a three-entry list, and a tetrahedron a four-entry list. For convenience certain entities of the mesh can be further defined in the domains dictionary. Each key defines a domain and maps to another dictionary. This second-level dictionary contains at least two keys: "dimension" mapping to the dimension of the specified domain (1,2, or 3) and "simplices" containing a list of integers mapping into the respective simplex lists. More keys may be added to the dictionary to define additional and/or custom information on the domain. For instance the compute_size! function adds an entry with the domain size.

WavesAndEigenvalues.Meshutils.Mesh — Method

mesh=Mesh(file_name::String; scale=1)

read a tetrahedral mesh from gmsh file into mesh. The optional scaling factor scale may be used to scale the units of the mesh.

WavesAndEigenvalues.Meshutils.generate_field — Function

generate_field(mesh::Mesh,func,el_type=0)

Generate a field by applying the function func onto the mesh mesh. The element type is either el_type=0 for field values associated with the mesh tetrahedra or el_type=1 for field values associated with the mesh vertices. The function func must accept exactly three input numerical arguments, corresponding to the three space dimensions.

WavesAndEigenvalues.Meshutils.vtk_write — Function

vtk_write(file_name, mesh, data)

Write vtk-file containing the datavalues data given on the usntructured grid mesh.

Arguments

file_name::String: name given to the written files. The name will be preceeded by a specifier. See Notes below.

mesh::Mesh: mesh associated with the data.

data::Dict: Dictionairy containing the data.

Notes

The routine automatically sorts the data according to its type and writes it in up to three diffrent files. Data that is constant on a tetrahedron goes into "$(filename)_const.vtu", data that is linearly interpolated on a tetrahedron goes into "$(filename)_lin.vtu", and data that is quadratically interpolated on a tetrahedron goes into "$(filename)_quad.vtu"

WavesAndEigenvalues.Meshutils.extend_mesh — Function

full_mesh=extend_mesh(mesh::Mesh, doms; sym_name="Symmetry", blch_name="Bloch", unit=false)

Create full mesh or unit cell from half cell represented in mesh.

Arguments

mesh::Mesh: mesh representing the half cell. The mesh must span a sector of 2π/2N, where N is the (integer) degree of symmetry of the full mesh.

doms::List: list of 2-tuples containing the domain names and their copy_degree (see notes below).

sym_name::String="Symmetry": name of the domain of mesh that forms the symmetry plane of the half-mesh.

blch_name::String="Bloch": name of the domain of mesh that forms the remaining azimuthal plane of the half-mesh.

unit::Bool=false: toggle whether extend the mesh to unit cell only.

Returns

full_mesh::Mesh: representation of the full mesh.

Notes

The routine copies only domains that are specified in doms. These domains are extended according to the specified copy_degree. The following are available:

:full: extent the domain and save it under the same name in full_mesh.

:unit: extent the domain and save the individual unit cells labeled from 0 to N-1 in full_mesh.

:half: extent the domain and save the individual unit cells as half cells labeled from 0 to N-1 in full_mesh where one half-cell contains _img in its name.

Multiple styles can be mixed in one domain specification. An example for doms would be doms=[("Interior", :full), ("Outlet", :unit), ("Flame", :half)]

WavesAndEigenvalues.Meshutils.octosplit — Function

new_mesh=octosplit(mesh::Mesh)

Subdivide each tetrahedron in the mesh mesh into 8 tetrahedra by splitting each edge at its center.

Notes

The algorithm introduces length(mesh.lines) new vertices. This yields a finer mesh featuring 8*size(mesh,2) tetrahedra, 4*length(mesh.triangles) triangles, and 2*length(mesh.lines) lines. From the 3 possible subdivision of a tetrahedron the algorithm automatically chooses the one that minimizes the edge lengths of the new tetrahedra. The point labeling of new_mesh is consistent with the point labeling in mesh, i.e., the first size(mesh.points,2) points in new_mesh.points are identical to the points in mesh.points. Hence, mesh and new_mesh form a hierachy of meshes.