Mesh

WavesAndEigenvalues.Meshutils — Module

Module containing functionality to read and process tetrahedral meshes from gmsh.

WavesAndEigenvalues.Meshutils.Mesh — Type

Definition of the mesh type

Fields

name::String: the name of the mesh.

points::Array: 3×N array containing the coordinates of the N points defining the mesh.

lines::List: List of simplices defining the edges of the mesh

triangles::List: List of simplices defining the surface triangles of the mesh

tetrahedra::List: List of simplices defining the tetrahedra of the mesh

domains::Dict: Dictionairy defining the domains of the mesh. See comments below.

file::String: path to the file containing the mesh.

tri2tet::Array: Array of length length(tetrahedra) containing the indices of the connected tetrahedra.

dos: special field meant to contain symmetry information of highly symmetric meshes.

Notes

The meshes are supposed to be tetrahedral. All simplices (lines, triangles, and tetrahedra) are stored as list of simplices. Simplices are lists of integers containing the indices of the points (i.e. the column number in the points array) forming the simplex. This means a line is a two-entry list, a triangle a three-entry list, and a tetrahedron a four entry lists. For convenience certain entities of the mesh can be further defined in the domainsdictionary. Each key defines a domain and maps to another dictionary. This second-level dictionary contains at least two keys: "dimension" mapping to the dimension of the specified domain (1,2, or 3) and "simplices" containing a list of integers mapping into the respective simplex lists. More keys may be added to the dictionary to define additional and/or custom information on the domain. For instance the compute_volume! function adds an entry with the domain size.

WavesAndEigenvalues.Meshutils.Mesh — Method

mesh=Mesh(file_name::String; scale=1)

read a tetrahedral mesh from gmsh file into mesh. The optional scaling factor scale may be used to scale the units of the mesh.

WavesAndEigenvalues.Meshutils.collect_lines! — Method

collect_lines!(mesh::Mesh)

Populate the list of lines in mesh.lines.

WavesAndEigenvalues.Meshutils.compute_volume! — Method

compute_volume!(mesh::Mesh,dom::String)

Compute the size of the domain dom and store it in the field "size" of the domain definition. The size will be a length, area, or volume depending on the dimension of the domain.

WavesAndEigenvalues.Meshutils.create_full_mesh — Method

full_mesh=create_full_mesh(mesh::Mesh, doms; sym_name="Symmetry", blch_name="Bloch", unit=false)

Create full mesh from half cell represented in mesh.

Arguments

mesh::Mesh: mesh representing the half cell. The mesh must span a sector of 2π/2N, where N is the (integer) degree of symmetry of the full mesh.

doms::List: list of 2-tuples containing the domain names and their copy_degree (see notes below).

sym_name::String="Symmetry": name of the domain of mesh that forms the symmetry plane of the half-mesh.

blch_name::String="Bloch": name of the domain of mesh that forms the remaining azimuthal plane of the half-mesh.

unit::Bool=false: toggle whether extend the mesh to unit cell only.

Returns

full_mesh::Mesh: representation of the full mesh.

Notes

The routine copies only domains that are specified in doms. These domains are extended according to the specified copy_degree. The following are available:

:full: extent the domain and save it under the same name in full_mesh.

:unit: extent the domain and save the individual unit cells labeled from 0 to N-1 in full_mesh.#

:half: extent the domain and save the individual unit cells as half cells labeled from 0 to N-1 in fullmesh where one half-cell contains `img` in its name.

Multiple styles can be mixed in one domain specification. An example for doms would be doms=[("Interior", :full), ("Outlet", :unit), ("Flame", :half)]

WavesAndEigenvalues.Meshutils.find_tetrahedron_containing_point — Method

tet_idx=find_tetrahedron_containing_point(mesh::Mesh,point)

Find the tetrahedron containing the point point and return the index of the tetrahedron as tet_idx. If the point lies on the interface of two or more tetrahedra, the returned tet_idx will be the lowest index of these tetrahedra, i.e. the index depends on the ordering of the tetrahedra in the mesh.tetrahedra. If no tetrahedron in the mesh encloses the point tet_idx==0.

WavesAndEigenvalues.Meshutils.link_triangles_to_tetrahedra! — Method

link_triangles_to_tetrahedra!(mesh::Mesh)

Find the tetrahedra that are connected to the triangles in mesh.triangles and store this information in mesh.tri2tet.

WavesAndEigenvalues.Meshutils.unify! — Method

unify!(mesh::Mesh ,new_domain::String,domains...)

Create union of domains and name it new_domain. domains must share the same dimension and new_domain must be a new name otherwise errors will occur.

WavesAndEigenvalues.Meshutils.compare — Method

cmpr=compare(A,B)

Compare two simplices following lexicographic rules. If B has lower rank, then A cmpr ==-1. If B has higher rank than A, cmpr==1. If A and B are identical, cmpr==0.

WavesAndEigenvalues.Meshutils.create_rotation_matrix_around_axis — Method

R = create_rotation_matrix_around_axis(n,α)

Compute the rotation matrix around the directional vector nrotating by an angle α (in rad).

WavesAndEigenvalues.Meshutils.find_foot_of_perpendicular — Method

foot=find_foot_of_perpendicular(pnt,pln)

Compute the position of the foot of the perpendicular from the point pnt to the plane pln.

#Notes Code adapted from https://www.geeksforgeeks.org/mirror-of-a-point-through-a-3-d-plane/

WavesAndEigenvalues.Meshutils.find_intersection_of_two_planes — Method

p,n = find_intersection_of_two_planes(pln1,pln2)

Find a parametrization of the axis defined by the intersection of the planes pln1and pln2. The axis is parameterized by a point p and direction n.

#Notes The algorithm follows John Krumm's solution for finding a common point on two intersecting planes as it is explained in https://math.stackexchange.com/questions/475953/how-to-calculate-the-intersection-of-two-planes To simplify the code, here the point p0 is the origin.

WavesAndEigenvalues.Meshutils.find_smplx — Method

idx = find_smplx(list,smplx)

Find index of simplex in ordered list of simplices. If the simplex is not contained in the list, the returned index is nothing

WavesAndEigenvalues.Meshutils.insert_smplx! — Method

insert_smplx!(list,smplx)

Insert simplex in ordered list of simplices.

WavesAndEigenvalues.Meshutils.is_point_in_plane — Method

check = is_point_in_plane(pnt, pln)

Check whether point "pnt" is in plane . Note that there is currently no handling of round off errors.

WavesAndEigenvalues.Meshutils.make_normal_outwards! — Method

makenormaloutwards!(pln,testpoint)

Ensure that the parameterization of the plane "pln" has a normal that is directed twoards testpoint. If necessary reparametrize pln.

WavesAndEigenvalues.Meshutils.read_ansys — Method

read a mesh from ansys cfx file

WavesAndEigenvalues.Meshutils.read_msh2 — Method

points, lines, triangles, tetrahedra, domains=read_msh2(file_name)

Read gmsh's .msh-format version 2. This is an old format and wherever possible version 4 should be used.

http://gmsh.info/doc/texinfo/gmsh.html#MSH-file-format

WavesAndEigenvalues.Meshutils.read_msh4 — Method

points, lines, triangles, tetrahedra, domains=read_msh4(file_name)

Read gmsh's .msh-format version 4.

http://gmsh.info/doc/texinfo/gmsh.html#MSH-file-format

WavesAndEigenvalues.Meshutils.reflect_point_at_plane — Method

p=reflect_point_at_plane(pnt,pln)

Reflect point pnt at plane pln and return as p.

#Notes Code adapted from https://www.geeksforgeeks.org/mirror-of-a-point-through-a-3-d-plane/

WavesAndEigenvalues.Meshutils.sort_smplx — Method

sort_smplx(list,smplx)

Helper function for sorting simplices in lexicographic manner. Utilize divide an conquer strategy to find a simplex in ordered list of simpleces.

WavesAndEigenvalues.Meshutils.three_points_to_plane — Method

pln=three_points_to_plane(A)

Compute equation for a plane from the three points defined in A. The plane is parameterized as a*x+b*y+c*z+d==0 with the coefficients returned as pln=[a,b,c,d].

#Arguments

A::3×3-Array : Array containing the three points defining the plane as columns.

#Notes Code adapted from https://www.geeksforgeeks.org/program-to-find-equation-of-a-plane-passing-through-3-points/