Helmholtz

WavesAndEigenvalues.Helmholtz — Module

Module providing functionality to numerically discretize the (thermoacoustic) Helmholtz equation by first and second-order finite elements.

WavesAndEigenvalues.Helmholtz.bloch_expand — Function

v=bloch_expand(mesh::Mesh,sol::Solution,b=:b)

Expand solution vector sol.v on mesh meshwith Bloch wave number sol.params[b] and return it as v.

WavesAndEigenvalues.Helmholtz.discretize — Method

L=discretize(mesh, dscrp, C; el_type=1, c_type=0, b=:__none__, mass_weighting=true)

Discretize the Helmholtz equation using the mesh mesh.

Arguments

mesh::Mesh: tetrahedral mesh

dscrp::Dict: dictionary containing information on where to apply physical constraints. Such as standard wave propagation, boundary conditions, flame responses, etc.

C:Array: array defining the speed of sound. If c_type==0 the speed of sound is constant along one tetrahedron and length(C)==length(mesh.tetrahedra). If c_type==1 the speed of sound is linearly interpolated between the vertices of the mesh and length(C)==size(mesh.points,2).

el_type = 1: optional paramater to select between first (el_type==1 the default) and second order (el_type==2) finite elements.

c_type = 1: optional parameter controlling whether speed of sound is constant on a tetrahedron or linearly interpolated between vertices.

b::Symbol=:__none__: optional parameter defining the Bloch wave number. If b=:__none__ (the default) no Blochwave formalism is applied.

mass_weighting=true: optional parameter if true mass matrix is used as weighting matrix for householder, otherwise this matrix is not set.

Returns

L::LinearOperatorFamily: parametereized discretization of the specified Helmholtz equation.

Discretization Order

WavesAndEigenvalues.Helmholtz.lin.assemble_boundary_mass_operator — Method

boundary mass operator for first order elements

WavesAndEigenvalues.Helmholtz.lin.assemble_stiffness_operator — Method

stiffness operator for first order elements

WavesAndEigenvalues.Helmholtz.lin.assemble_volume_source — Method

volume source vector for first order elements