NLEVP

WavesAndEigenvalues.NLEVP — Module

Module containing routines to solve and perturb nonlinear eigenvalue problems.

WavesAndEigenvalues.NLEVP.LinearOperatorFamily — Method

L=LinearOperatorFamily(fname::String)

Load and construct LinearOperatorFamily from file fname.

See also: save

WavesAndEigenvalues.NLEVP.beyn — Method

Ω, P = beyn(L::LinearOperatorFamily, Γ; <keyword arguments>)

Compute all eigenvalues of L inside the contour Γ together with the associated eigenvectors. The contour is given as a list of complex numbers which are interpreted as polygon vertices in the complex plane. The eigenvalues are stored in the list Ω and the eigenvectors are stored in the columns of P. The eigenvector P[:,i] corresponds to the eigenvalue Ω[i], i.e., they satisfy L(Ω[i])P[:,i]=0.

Arguments

L::LinearOperatorFamily: Definition of the non-linear eigenvalue problem

Γ::Array: List of complex points defining the contour.

l::Integer = 5: estimate of the number of eigenvalues inside of Γ.

K::Integer = 1: Augmention dimension if Γ is assumed to contain more than size(L)[1] eigenvalues.

N::Integer = 16: Number of evaluation points used to perform the Gauss-Legendre integration along each edge of Γ.

tol::Float=0.0: Threshold value to discard spurious singular values. If set to 0 (the default) no singular values are discarded.

pos_test::bool=true: If set to true perform positions test on the computed eigenvalues, i.e., check whether the eigenvalues are enclosed by Γ and disregard all eigenvalues which fail the test.

output::bool=false: Show progressbar if true.

Returns

Ω::Array: List of computed eigenvalues

P::Matrix: Matrix of eigenvectors.

Notes

The original algorithm was first presented by Beyn in [1]. The implementation closely follows the pseudocode from Buschmann et al. in [2].

References

[1] W.-J. Beyn, An integral method for solving nonlinear eigenvalue problems, Linear Algebra and its Applications, 2012, 436(10), p.3839-3863, https://doi.org/10.1016/j.laa.2011.03.030

[2] P.E. Buschmann, G.A. Mensah, J.P. Moeck, Solution of Thermoacoustic Eigenvalue Problems with a Non-Iterative Method, J. Eng. Gas Turbines Power, Mar 2020, 142(3): 031022 (11 pages) https://doi.org/10.1115/1.4045076

See also: householder

WavesAndEigenvalues.NLEVP.householder — Method

sol::Solution, n, flag = householder(L::LinearOperatorFamily, z; <keyword arguments>)

Use a Householder method to iteratively find an eigenpar of L, starting the the iteration from z.

Arguments

L::LinearOperatorFamily: Definition of the nonlinear eigenvalue problem.

z: Initial guess for the eigenvalue.

maxiter::Integer=10: Maximum number of iterations.

tol=0: Absolute tolerance to trigger the stopping of the iteration. If the difference of two consecutive iterates is abs(z0-z1)<tol the iteration is aborted.

relax=1: relaxation parameter

lam_tol=0.: tolerance for the auxiliary eigenvalue to test convergence

order::Integer=1: Order of the Householder method, Maximum is 5

n_eig_val::Integer=1: Number of Eigenvalues to be searched for in intermediate ARPACK calls.

v0::Vector: Initial vector for Krylov subspace generation in ARPACK calls. If not provided the vector is initialized with ones.

output::Bool: Toggle printing online information.

Returns

sol::Solution

n::Integer: Number of perforemed iterations

flag::Integer: flag reporting the success of the method. Most important values are 1: method converged, 0: convergence might be slow, -1:maximum number of iteration has been reached. For other error codes see the source code.

Notes

Housholder methods are a generalization of Newton's method. If order=1 the Housholder method is identical to Newton's method. The solver then reduces to the "generalized Rayleigh Quotient iteration" presented in [1]. If no relaxation is used (relax == 1), the convergence rate is of order+1. With relaxation (relax != 1) the convergence rate is 1. Thus, with a higher order less iterations will be necessary. However, the computational time must not necessarily improve nor do the convergence properties. Anyway, if the method converges, the error in the eigenvalue is bounded above by tol. For more details on the solver, see the thesis [2].

References

[1] P. Lancaster, A Generalised Rayleigh Quotient Iteration for Lambda-Matrices,Arch. Rational Mech Anal., 1961, 8, p. 309-322, https://doi.org/10.1007/BF00277446

[2] G.A. Mensah, Efficient Computation of Thermoacoustic Modes, Ph.D. Thesis, TU Berlin, 2019

See also: beyn,

WavesAndEigenvalues.NLEVP.perturb! — Method

perturb!(sol::Solution,L::LinearOperatorFamily,param::Symbol,N::Int; <keyword arguments>)

Compute the Nth order power series perturbation coefficients for the solution sol of the nonlineaer eigenvalue problem given by the operator family L with respect to the parameter param. The coefficients will be stored in the field sol.eigval_pert and sol.v_pert for the eigenvalue and the eignevector, respectively.

Keyword Arguments

mode = :compact: parameter controlling internal programm flow. Use the default, unless you know what you are doing.

Notes

For large perturbation orders N the method might be slow.

See also: perturb_fast!

WavesAndEigenvalues.NLEVP.perturb_fast! — Method

perturb_fast!(sol::Solution,L::LinearOperatorFamily,param::Symbol,N::Int; <keyword arguments>)

Compute the Nth order power series perturbation coefficients for the solution sol of the nonlineaer eigenvalue problem given by the operator family L with respect to the parameter param. The coefficients will be stored in the field sol.eigval_pert and sol.v_pert for the eigenvalue and the eignevector, respectively.

Keyword Arguments

mode = :compact: parameter controlling internal programm flow. Use the default, unless you know what you are doing.

Notes

This method reads multi-indeces for the computation of the power series coefficients from disk. Make sure that JulHoltz is properly installed to use this fast method.

See also: perturb!

WavesAndEigenvalues.NLEVP.perturb_norm! — Method

perturb_norm!(sol::Solution,L::LinearOperatorFamily,param::Symbol,N::Int; <keyword arguments>)

Compute the Nth order power series perturbation coefficients for the solution sol of the nonlineaer eigenvalue problem given by the operator family L with respect to the parameter param. The coefficients will be stored in the field sol.eigval_pert and sol.v_pert for the eigenvalue and the eignevector, respectively.

Keyword Arguments

mode = :compact: parameter controlling internal programm flow. Use the default, unless you know what you are doing.

Notes

This method reads multi-indeces for the computation of the power series coefficients from disk. Make sure that JulHoltz is properly installed to use this fast method.

See also: perturb!

WavesAndEigenvalues.NLEVP.save — Method

save(fname::String,L::LinearOperatorFamily)

Save L to file fname. The file is utf8 encoded and adheres to a Julia-enriched TOML standard.

See also: LinearOperatorFamily

WavesAndEigenvalues.NLEVP.polyval — Method

`f=polyval(p,z)`

Evaluate polynomial f(z)=∑_i p[i]z^1 at z, using Horner's scheme.

WavesAndEigenvalues.NLEVP.read_toml — Method

This is a minimal TOML parser. See https://en.wikipedia.org/wiki/TOML